requirements.txt
# 基础环境（Python3.14+）
# 异步采集/接口（IO密集型核心）
pyzmq>=25.0.0
aiohttp>=3.9.1
websockets>=12.0
# 数据处理（期货行情时间序列核心）
pandas>=2.1.0
numpy>=1.26.0
# 配置解析（yaml易读，适配复杂配置）
PyYAML>=6.0.1
# CTP/C++接口Python封装
cffi>=1.16.0
# 时序数据库/缓存（可选，根据存储方案调整）
influxdb-client>=1.39.0
redis>=5.0.1
# 测试
pytest>=7.4.3
# 可视化（可选，异常检测/回测用）
plotly>=5.17.0
matplotlib>=3.8.0
# 其他工具（装饰器/数据校验）
decorator>=5.1.1
pydantic>=2.4.2
pybind11>=3.0.0